Refactor file structure for commands test files
